move SimpleExpansionPanel to common utilities
