+++ src/components/subscriptions/edit/EditQuotas.js
@@ -4,7 +4,7 @@ import { FieldArray, Form, Formik } from "formik";
 import { mapPropsToValues, formatQuotas } from "./formatters";
 import DEDialog from "components/utils/DEDialog";
 import { Button, Grid, Typography } from "@material-ui/core";
-import SimpleExpansionPanel from "components/tools/SimpleExpansionPanel";
+import SimpleExpansionPanel from "components/utils/SimpleExpansionPanel";
 
 import {
     updateUserQuotas,
