+++ src/components/tools/edit/ContainerPorts.js
@@ -4,7 +4,7 @@ import { useTranslation } from "i18n";
 
 import { DeleteBtn, AddBtn } from "../Buttons";
 import ids from "../ids";
-import SimpleExpansionPanel from "../SimpleExpansionPanel";
+import SimpleExpansionPanel from "components/utils/SimpleExpansionPanel";
 import { minValue, nonEmptyMinValue } from "components/utils/validations";
 import { DERow } from "components/table/DERow";
 import DETableHead from "components/table/DETableHead";
