+++ src/components/tools/edit/ToolRestrictions.js
@@ -2,7 +2,7 @@ import React from "react";
 import { useTranslation } from "i18n";
 
 import ids from "../ids";
-import SimpleExpansionPanel from "../SimpleExpansionPanel";
+import SimpleExpansionPanel from "components/utils/SimpleExpansionPanel";
 import globalConstants from "../../../constants";
 
 import buildID from "components/utils/DebugIDUtil";
